feat: refactor underlying extraction logic and add utility function for shortcode parsing
