+++ packages/components/src/components/positions-drawer-card/positions-drawer-card.tsx
@@ -15,6 +15,7 @@ import {
     isVanillaContract,
 } from '@deriv/shared';
 import { TContractInfo, TContractStore } from '@deriv/shared/src/utils/contract/contract-types';
+import { getUnderlyingFromShortcode } from '../../utils/contract-helpers';
 import { TToastConfig } from '../types/contract.types';
 
 type TPositionsDrawerCardProps = {
@@ -153,24 +154,7 @@ const PositionsDrawerCard = ({
         </React.Fragment>
     );
 
-    // Compute effective underlying with fallback extraction from shortcode
-    const getEffectiveUnderlying = () => {
-        if (contract_info?.underlying) {
-            return contract_info.underlying;
-        }
-
-        // Fallback: Extract underlying from shortcode if missing
-        if (contract_info?.shortcode) {
-            const parts = contract_info.shortcode.split('_');
-            if (parts.length >= 2) {
-                return parts[1];
-            }
-        }
-
-        return null;
-    };
-
-    const effective_underlying = getEffectiveUnderlying();
+    const effective_underlying = contract_info?.underlying || getUnderlyingFromShortcode(contract_info?.shortcode);
 
     const contract_card_body = is_link_disabled ? (
         <div className={contract_card_classname}>{effective_underlying ? contract_el : loader_el}</div>
